(ref) remove extra build_shared.py
